requirements-prod.txt
absl-py==2.3.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
astunparse==1.6.3
attrs==25.4.0
boto3==1.42.39
botocore==1.42.39
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
dateparser==1.2.2
flatbuffers==25.12.19
frozenlist==1.8.0
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
h5py==3.15.1
idna==3.11
infisicalsdk==1.0.15
jmespath==1.1.0
keras==3.13.0
libclang==18.1.1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.1
namex==0.1.0
numba==0.61.2
numpy==2.2.6
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pycparser==3.0
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==1.10.26
Pygments==2.19.2
PyNaCl==1.6.2
pyproject_hooks==1.2.0
python-binance==1.0.34
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.46
ta==0.11.0
TA-Lib==0.6.8
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.20.0
termcolor==3.2.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
websockets==16.0
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
yarl==1.22.0
